+++ backend/db/300_create_course_data_by_tee.sh
@@ -2,6 +2,9 @@
 #!/bin/bash
 set -e
 
+source ${HOME}/sites/vhs/.env
+echo "DB_CONTAINER: $DB_CONTAINER"
+
 # Check if container is running
 if ! docker ps | grep -q $DB_CONTAINER; then
     echo "Error: Database container '$DB_CONTAINER' is not running"