+++ backend/db/300_create_course_names.sh
@@ -1,6 +1,7 @@
-
 #!/bin/bash
 set -e
+
+source ${HOME}/sites/vhs/.env
 
 # Check if container is running
 if ! docker ps | grep -q $DB_CONTAINER; then